[FIX] sign_oca Fix who can sign
Before this commit
------------------

Internal user cannot sign a document if this user's partner is a child of some other partner.

Example:
User "Pepe Frog" has a partner "Meme Inc, Pepe Frog" set as `user.partner_id`.
This is because partner "Meme Inc" is set as a company for the partner "Pepe Frog".
So when user Pepe Frog is assigned to a sign request he cannot see the "Sign" button.

After this commit
-----------------

User can see the "Sign" button and sign documents even if related partner has a parent partner set.